pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "spacy>=3.8.5",
    "torch>=2.6.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "fugashi[unidic]>=1.4.0",
    "hanlp>=2.1.1",
    "passlib>=1.7.4",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.0",
    "python-mecab-ko>=1.3.7",
    "qdrant-client>=1.13.3",
    "sqlalchemy>=2.0.40",
    "transformers>=4.50.2",
    "underthesea>=6.8.4",
    "uvicorn>=0.34.0",
]
